Move read_xml_tag to Support::XML
